refactor(control): extract runControlActionRoute into control-route module
runControlActionRoute is a generic control-route helper used by the
non-component pause/play route, so it does not belong in a module named
"components". Move it (and ControlRouteResponse) into control-route.ts and
update the factory and pause/play route imports. No behaviour change.
