+++ src/server/components.ts
@@ -1,15 +1,11 @@
-import {
-    ApiResponse,
-    Component,
-    ComponentType,
-    ControlAction,
-    Game,
-} from "@fc/types/types";
+import { Component, ComponentType, Game } from "@fc/types/types";
 import { Either, isLeft } from "fp-ts/Either";
 import { MakeLeft, MakeRight } from "@fc/lib/io-ts-helpers";
 import { NextRequest, NextResponse } from "next/server";
-import { getGameRepo, UPDATE_CONFLICT } from "@fc/server/repository/game";
-import { toApiResponse } from "@fc/server/turn";
+import {
+    ControlRouteResponse,
+    runControlActionRoute,
+} from "@fc/server/control-route";
 
 /**
  * The concrete component variant for a given componentType discriminant.
@@ -126,64 +122,10 @@ export function componentAction<T extends ComponentType, B>(
     };
 }
 
-type ControlRouteResponse = NextResponse<ApiResponse | { error: string }>;
-
 type ComponentRouteHandler = (
     request: NextRequest,
     props: { params: Promise<{ id: string }> },
 ) => Promise<ControlRouteResponse>;
-
-/**
- * Runs a control action for a game and maps the result to a response, shared by
- * the component routes and the pause/play control route.
- *
- * The repository update is a compare-and-set on turnInformation, which can lose
- * to a concurrent writer (e.g. a lazy auto-advance triggered by a player's
- * poll). On such a conflict this re-fetches the game and re-applies the action
- * once; a persistent conflict surfaces as HTTP 409 rather than a silent success
- * with the other writer's state. See #783.
- */
-export async function runControlActionRoute(
-    id: string,
-    action: ControlAction,
-): Promise<ControlRouteResponse> {
-    const gameRepo = getGameRepo();
-
-    if (isLeft(gameRepo)) {
-        return NextResponse.json({ error: gameRepo.left }, { status: 500 });
-    }
-
-    const game = await gameRepo.right.get(id);
-
-    if (isLeft(game)) {
-        return NextResponse.json({ error: "Game not found" }, { status: 404 });
-    }
-
-    let result = await gameRepo.right.runControlAction(game.right, action);
-
-    if (isLeft(result) && result.left === UPDATE_CONFLICT) {
-        const fresh = await gameRepo.right.get(id);
-
-        if (isLeft(fresh)) {
-            return NextResponse.json(
-                { error: "Game not found" },
-                { status: 404 },
-            );
-        }
-
-        result = await gameRepo.right.runControlAction(fresh.right, action);
-    }
-
-    if (isLeft(result)) {
-        if (result.left === UPDATE_CONFLICT) {
-            return NextResponse.json({ error: "conflict" }, { status: 409 });
-        }
-
-        return NextResponse.json({ error: result.left }, { status: 500 });
-    }
-
-    return NextResponse.json(toApiResponse(result.right));
-}
 
 /**
  * Builds a control route handler for a component. Collapses the ~70-line
